% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overview.R
\name{check_numeric}
\alias{check_numeric}
\title{Check numeric}
\usage{
check_numeric(dataset)
}
\arguments{
\item{dataset}{Either a dataframe or element of a list}
}
\value{
\itemize{
  \item{name a}{description}
  \item B - A vector of numbers.
}
}
\description{
Basic statistics for numeric data
}
\examples{
For dataframe:
check_numeric(dataset)
For element of a list:
check_numeric(my_data$dataset)

}
