% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_datasets.R
\name{connect_datasets}
\alias{connect_datasets}
\title{Connect datasets}
\usage{
connect_datasets(..., remove_original_data = TRUE)
}
\arguments{
\item{...}{names of datasets that should be concatenated}

\item{remove_original_data}{If set to TRUE [default], it will remove all dataframes passed as arguments.}
}
\value{
Named list; dataset names are further used as a list's elements names. Common ID must be placed as a first column for all data.
}
\description{
Connect datasets
}
\examples{
my_data <-connect_datasets(clinical_data, eeg_data, proteomics_data, miRNA_data, RNA_data)

}
