% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{convert2matrix}
\alias{convert2matrix}
\title{Plot PCA}
\usage{
convert2matrix(x)
}
\arguments{
\item{data_to_plot}{data for PCA analysis}

\item{variable_to_color}{name of variable for grouping; must be categorical}

\item{plot_title}{}
}
\value{
None
}
\description{
Plot PCA
}
\examples{
data_to_plot <- data$RNA_data \%>\% left_join(data$clinical_data)
plot_PCA(data_to_plot, variable_to_color = sex, legend_title = "Sex", title = "RNA")

}
