% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_per_model.R
\name{count_stats_per_model}
\alias{count_stats_per_model}
\title{Count Statistics per Model}
\usage{
count_stats_per_model(data, target)
}
\arguments{
\item{data}{A dataframe containing the data for which statistics should be calculated.}

\item{target}{A list containing the target variable details. This list should contain:
`$target_variable` which specifies the target variable name as a string.}
}
\value{
A dataframe with calculated statistics per target level.
  - For numeric variables: Median, Q1, Q3 values.
  - For factor variables: Level percentages and counts.
}
\description{
This function calculates statistics for the given dataset per target level.
It first detects binary columns and converts them to factors.
Then, for numeric columns, it calculates the median, first quartile, and third quartile.
For factor columns, it calculates percentages and counts per factor level.
}
\examples{
\dontrun{
df <- data.frame(target_var = c("A", "A", "B", "B"),
                 num_var = c(1, 2, 2.5, 3),
                 fct_var = c("Yes", "No", "Yes", "No"))
target_list <- list(target_variable = "target_var")
count_stats_per_model(df, target_list)
}

}
