% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overview.R
\name{data_summary}
\alias{data_summary}
\title{Summarise data structure}
\usage{
data_summary(data)
}
\arguments{
\item{data}{Named list of dataframes}
}
\value{
Summary of data in the tabularized format with following positions:
number of samples, number of variables, number of numeric/character/factor columns across dataset.
}
\description{
Check the data structure for each element of list separately
}
\examples{
data1 <- data.frame(SampleID = c(1:5),Variable1 = c(3, 6, 2, 8, 4),
Variable2 = c("A", "B", "A", "C", "B"),Variable3 = factor(c("X", "Y", "Z", "X", "Y")))
data2 <- data.frame(SampleID = c(1:5), Variable1 = c(2, 4, 6, 8, 10),
Variable2 = c("A", "B", "C", "D", "E"),Variable3 = factor(c("X", "Y", "Z", "X", "Y")))
my_data <- list(data1 = data1, data2 = data2)
data_summary(my_data)

}
