% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_target.R
\name{define_target}
\alias{define_target}
\title{Define a target for a classification problem}
\usage{
define_target(
  phenotype_df_name,
  id_variable_name,
  target_variable_name,
  positive_class_name = NULL
)
}
\arguments{
\item{phenotype_df_name}{Name of a dataframe containing the phenotype/clinical data.}

\item{id_variable_name}{Name of the ID variable in the phenotype data frame.}

\item{target_variable_name}{Name of the target variable in the phenotype data frame.}

\item{positive_class_name}{Name of the positive class indication in the target variable.}
}
\value{
A list containing the most important information about defined target
}
\description{
Given the names of the input data frame, id variable, target variable,
and positive class variable, this function creates a list containing
these objects for use in a classification task.
}
\examples{
my_target <- define_target(phenotype_df_name = "clinical_data", id_variable_name = "patient_id",
    target_variable_name = "survived", positive_class_name = "1")

}
