% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefiltering.R
\name{filter_below_threshold}
\alias{filter_below_threshold}
\title{Filter values below defined threshold}
\usage{
filter_below_threshold(
  data,
  numeric_threshold,
  pcent_of_samples,
  inside_group = NULL,
  group_by = NULL
)
}
\arguments{
\item{numeric_threshold}{value below which data should be filtered out}

\item{pcent_of_samples}{indicates % of values within variable which should contain
a value higher than the one defined in the numeric_variable parameter; n total is calculated based
on all samples (no distinction between missing and non-missing values)}
}
\description{
Filter values below defined threshold
}
