% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefiltering.R
\name{filter_missing}
\alias{filter_missing}
\title{Filter missing values}
\usage{
filter_missing(data, pcent_of_samples, inside_group = NULL)
}
\arguments{
\item{pcent_of_samples}{indicates % of values within variable which should contain
non-missing values in all samples}
}
\description{
Filter missing values
}
