% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mlflow_data.R
\name{get_metrics_for_all_data}
\alias{get_metrics_for_all_data}
\title{Get experiment metrics}
\usage{
get_metrics_for_all_data(experiment_name, n_cores = detectCores()/4)
}
\description{
Get experiment metrics
}
