% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_and_explain.R
\name{load_and_explain}
\alias{load_and_explain}
\title{Load Explainer and Generate Explanations}
\usage{
load_and_explain(path, data)
}
\arguments{
\item{path}{The directory path where the explainer is saved.}

\item{data}{The data for which explanations are needed.}
}
\value{
The explainer's output for the data. If there is an error, the function logs the error message and returns NULL.
}
\description{
This function loads a saved explainer from a specified path and generates explanations
for the provided data.
}
\examples{
\dontrun{
explanations <- load_and_explain("/path/to/explainer/directory", sample_data)
}

}
