% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_and_explain.R
\name{load_and_predict}
\alias{load_and_predict}
\title{Load Model and Make Predictions}
\usage{
load_and_predict(path, data)
}
\arguments{
\item{path}{The directory path where the model is saved.}

\item{data}{The data on which predictions should be made.}
}
\value{
The model's predictions on the data. If there is an error, the function logs the error message and returns NULL.
}
\description{
This function loads a saved model from a specified path and makes predictions
on the provided data.
}
\examples{
\dontrun{
predicted_values <- load_and_predict("/path/to/model/directory", sample_data)
}

}
