% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_multiple.R
\name{model_multiple}
\alias{model_multiple}
\title{Create small models}
\usage{
model_multiple(
  data,
  target,
  experiment_name,
  n_cores = parallel::detectCores()/4,
  n_prop = 2/3,
  n_repeats = 50,
  directory = getwd()
)
}
\description{
Create small models
}
