% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coverage.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Visualize intersections between different modalities/}
\usage{
plot_coverage(data)
}
\arguments{
\item{data}{named list of connected dataframes. Common ID must be placed as a first column for all data.}
}
\value{
plot presenting data coverage between different datasets
}
\description{
Visualize intersections between different modalities/
}
\examples{
plot_coverage(my_data)

}
