% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overview.R
\name{plot_density_numeric}
\alias{plot_density_numeric}
\title{Density plots for numeric data}
\usage{
plot_density_numeric(dataset)
}
\arguments{
\item{dataset}{Either a dataframe or element of a list}
}
\description{
Basic statistics for numeric data
}
\examples{
For dataframe:
check_numeric(dataset)
For element of a list:
check_numeric(my_data$dataset)

}
