% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overview.R
\name{plot_stats}
\alias{plot_stats}
\title{Plot statistics for numeric data}
\usage{
plot_stats(dataset, metric_to_plot)
}
\arguments{
\item{dataset}{Either a dataframe or element of a list}

\item{metric_to_plot}{The metric to plot (e.g., 'median', 'mean', etc.).}
}
\value{
A boxplot and a table summarizing statistics for the specified metric.
}
\description{
This function generates boxplots to visualize basic statistics for numeric data within a dataframe
}
\examples{
\dontrun{
# For a dataframe:
plot_stats(my_data, "median")
# For an element of a list:
plot_stats(my_data$dataset, "mean")
}

}
