% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_for_modeling.R
\name{prepare_data_for_modelling}
\alias{prepare_data_for_modelling}
\title{Prepare data for modelling}
\usage{
prepare_data_for_modelling(
  data,
  target,
  remove_correlated_features = FALSE,
  normalize_data = FALSE
)
}
\arguments{
\item{data}{A named list of data.frames containing the data to be prepared for modelling.}

\item{target}{A named list with the following elements:
* phenotype_df: A character string specifying the dataframe containing the target variable.
* id_variable: A character string specifying the ID variable.
* target_variable: A character string specifying the target variable.
* positive_class: (Optional) The value of the target variable representing the positive class.}

\item{remove_correlated_features}{A logical value indicating whether to remove correlated features (default is FALSE).}
}
\value{
A named list of processed dataframes.
}
\description{
This function preprocesses the data for modelling, including one-hot encoding
of non-numeric columns, translating logical columns into numbers, converting the ID variable to a character
and the target variable to a factor. It also adds the dataset name to
each variable to distinguish data and relevels the target variable to set the positive class.
Additionally, it removes highly correlated predictors (default).
}
\examples{
\dontrun{
data <- list(
  phenotype_df = data.frame(id = 1:100, target = sample(c(0, 1), 100, replace = TRUE),
      feature1 = rnorm(100)),
  another_df = data.frame(id = 1:100, col1 = rnorm(100),
      col2 = factor(sample(letters[1:3], 100, replace = TRUE)))
)

target <- list(
  phenotype_df = "phenotype_df",
  id_variable = "id",
  target_variable = "target",
  positive_class = 1
)

prepared_data <- prepare_data_for_modelling(data, target)
}
}
