% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_and_select_features.R
\name{rank_and_select_features}
\alias{rank_and_select_features}
\title{Rank and Select Features from Data}
\usage{
rank_and_select_features(
  data,
  target,
  filter_name = "auc",
  cutoff_method = "top_n",
  cutoff_treshold = 10,
  selection_type = "combined",
  return_ranking_list = FALSE,
  n_fold = 5,
  n_threads = 1
)
}
\arguments{
\item{data}{A data frame or list of data frames containing predictor variables. For separate selection, `data` should be a single data frame with columns tagged for different datasets.}

\item{target}{A list containing the following elements:
- `id_variable`: Character, the column name identifying unique observations.
- `target_variable`: Character, the name of the target variable for classification.
- `positive_class`: Character, the class to be considered positive for binary classification.
- `phenotype_df`: Character, optional, specifying the main dataset name when `selection_type = "separate"`.}

\item{filter_name}{Character, the name of the filter for ranking features (e.g., `"auc"`, `"information_gain"`, `"variance"`, `"mrmr"`, `"jmim"`).}

\item{cutoff_method}{Character, the method for selecting features. Options are `"top_n"`, `"percentage"`, or `"threshold"`.}

\item{cutoff_treshold}{Numeric, the threshold for feature selection based on the specified `cutoff_method`.}

\item{selection_type}{Character, specifies the selection mode. Options are `"combined"` for ranking across all data or `"separate"` for individual datasets.}

\item{return_ranking_list}{Logical, if `TRUE`, returns both the filtered data and ranking list.}

\item{n_fold}{Integer, the number of folds for cross-validation. Default is `5`.}

\item{n_threads}{Integer, number of threads to use for multi-threaded filters. Default is `1`.}
}
\value{
Returns either a data frame of selected features or, if `return_ranking_list = TRUE`, a list containing:
- `filtered_data`: The filtered data frame after feature selection.
- `ranking_list`: A list of feature rankings for each dataset.
}
\description{
The `rank_and_select_features()` function performs feature ranking and selection on a given dataset using various ranking filters.
The function supports either separate or combined selection for different datasets within `data`, enabling flexibility in ranking and selection.
}
\examples{
# Example usage
target <- list(id_variable = "ID", target_variable = "Outcome", positive_class = "Positive", phenotype_df = "main")
result <- rank_and_select_features(data = mydata, target = target, filter_name = "auc", cutoff_method = "top_n", cutoff_treshold = 10)

}
