% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_variables.R
\name{rank_features}
\alias{rank_features}
\title{Select variables}
\usage{
rank_features(data, target, filter_name, n_threads = 1)
}
\arguments{
\item{data}{Name of a dataframe name containing the phenotype/clinical data}

\item{target}{Name of a column with statuses}

\item{filter_name}{Name of a filter to be applied (https://mlr.mlr-org.com/articles/tutorial/filter_methods.html#current-methods)}

\item{n_threads}{Number of threads for feature selection (as default set to 1)}

\item{cutoff_method}{One of the following: top_n, percentage, threshold}

\item{cutoff_treshold}{Depending of a cutoff method, a number of features to be selected, % of variables to be selected, a threshold above which features are selected.}
}
\description{
Select variables related to a target. Features are selected based on ranking that is created depending of selected method.
Then, three methods of selection can be chosen:
}
