% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_data.R
\name{read_logged_data}
\alias{read_logged_data}
\title{Get experiment metrics from logged files}
\usage{
read_logged_data(experiment_name, file_type = "metrics", directory = getwd())
}
\arguments{
\item{experiment_name}{The name of the experiment.}

\item{file_type}{The type of file to read; either "metrics", "model_logs" or "params". Default is \code{metrics}.}

\item{directory}{The directory where the RDS files are stored. Default is the current working directory (\code{getwd()}).}
}
\value{
A data frame containing the combined experiment metrics.
}
\description{
This function reads data from files stored in a specified directory.
}
\examples{
\dontrun{
# Read experiment metrics from the "my_experiment" directory
metrics_data <- read_logged_data("my_experiment", file_type = "metrics")
}

}
