% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_data.R
\name{read_model_data}
\alias{read_model_data}
\title{Get experiment metrics from stored RDS files}
\usage{
read_model_data(experiment_name, directory = getwd())
}
\arguments{
\item{experiment_name}{The name of the experiment.}

\item{directory}{The directory where the RDS files are stored. Default is the current working directory (\code{getwd()}).}
}
\value{
A data frame containing the combined experiment metrics.
}
\description{
This function reads and combines metrics data from RDS files stored in a specified directory.
}
\examples{
# Read experiment metrics from the "my_experiment" directory
metrics_data <- read_model_data("my_experiment")

}
