% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data_into_train_test.R
\name{split_data_into_train_test}
\alias{split_data_into_train_test}
\title{Split Data into Training and Testing Sets}
\usage{
split_data_into_train_test(data, target, prop = 4/5)
}
\arguments{
\item{data}{A named list of dataframes to be split, where one of the dataframes should be named as the target$phenotype_df.}

\item{target}{A named list with elements 'phenotype_df', 'id_variable', and 'target_variable' (optional).
'phenotype_df' should match one of the names in the 'data' list, 'id_variable' should be a column in that dataframe,
and 'target_variable' should be a column for stratification (if required).}

\item{prop}{A numeric value between 0 and 1 representing the proportion of the data to be used for the training set (default is 4/5).}
}
\value{
A list with two named elements, 'train_data' and 'test_data', each containing the corresponding split dataframes.
}
\description{
This function splits the given data into training and testing sets using stratified sampling if a target variable is provided.
It also splits other dataframes in the list based on the same train and test ID subsets.
}
\examples{
# Test data preparation
sample_data <- list(
  phenotype_df = data.frame(
    id = 1:100,
    target_variable = sample(c(0, 1), 100, replace = TRUE),
    feature1 = rnorm(100),
    feature2 = rnorm(100)
  ),
  another_df = data.frame(
    id = 1:100,
    col1 = rnorm(100),
    col2 = rnorm(100)
  )
)
target <- list(
  phenotype_df = "phenotype_df",
  id_variable = "id",
  target_variable = "target_variable"
)
split_result <- split_data_into_train_test(sample_data, target, prop = 4/5)
}
