% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overview.R
\name{check_data}
\alias{check_data}
\title{Check numeric}
\usage{
check_data(dataset)
}
\arguments{
\item{dataset}{Either a dataframe or element of a list}
}
\value{
\itemize{
  \item A - The letters of the alphabet.
  \item B - A vector of numbers.
}
}
\description{
Basic statistics for numeric data
}
\examples{
#For dataframe:
check_numeric(dataset)
#For element of a list:
check_numeric(my_data$dataset)

}
