% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model.R
\name{create_model}
\alias{create_model}
\title{Create and evaluate a logistic regression model}
\usage{
create_model(
  train_data,
  test_data,
  target,
  validation_method = "cv",
  n_prop = 2/3,
  n_repeats = 5,
  log_experiment = TRUE,
  explain = TRUE,
  add_weights = FALSE,
  directory = getwd()
)
}
\arguments{
\item{train_data}{A dataframe containing the training dataset.}

\item{test_data}{A dataframe containing the test dataset.}

\item{target}{A named list with "id_variable" and "target_variable" specifying the ID and target variable names.}

\item{validation_method}{A character string specifying the validation method to be used; either "subsampling" or "cv" (default: "cv").}

\item{n_prop}{A numeric value representing the proportion of data used for each resample in the subsampling process (default: 2/3 for subsampling).}

\item{n_repeats}{A numeric value specifying the number of times to repeat the validation (default: 5).}

\item{log_experiment}{A logical value indicating whether to log the experiment details and results. Default is TRUE.}

\item{explain}{A logical value indicating whether to create model's explainer using DALEX (default: TRUE).}

\item{add_weights}{A logical value indicating whether to add weights to the training data based on the class imbalance. Default is FALSE.}

\item{directory}{A character string indicating the directory where experiment logs and models should be saved. Default is the current working directory.}
}
\value{
A tibble containing details of the model, training metrics, and test metrics (if test data provided).
In case of an error during processing, it returns a tibble with only the 'model_id'.
}
\description{
This function creates a logistic regression model using tidymodels, recipes, and workflows,
and evaluates its performance on training and test datasets. The function also logs the experiment
information, metrics, and artifacts (optional) and provides explanation with DALEX package.
}
