% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multiple_models.R
\name{create_multiple_models}
\alias{create_multiple_models}
\title{Create small models}
\usage{
create_multiple_models(
  data_filtered,
  experiment_name,
  n_cores = detectCores()/4,
  n_prop = 2/3,
  n_repeats = 50
)
}
\description{
Create small models
}
