% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_overview.R
\name{data_summary}
\alias{data_summary}
\title{Summarise data structure}
\usage{
data_summary(data)
}
\arguments{
\item{data}{Named list of dataframes}
}
\value{
Summary of data in the tabularized format with following positions:
number of samples, number of variables, number of numeric/character/factor columns across dataset.
}
\description{
Check the data structure for each element of list separately
}
\examples{
data_summary(my_data)

}
