% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_univariate_filtering.R
\name{nested_filtering}
\alias{nested_filtering}
\title{Univariate feature selection}
\usage{
nested_filtering(
  data,
  target,
  filter_name = "auc",
  cutoff_method = "top_n",
  cutoff_treshold = 10,
  n_threads = 1,
  nfold = 5
)
}
\arguments{
\item{data}{Name of a dataframe name containing the phenotype/clinical data}

\item{target}{Name of a column with statuses}

\item{filter_name}{Name of a filter to be applied (https://mlr.mlr-org.com/articles/tutorial/filter_methods.html#current-methods, column "Classif")}

\item{cutoff_method}{One of the following: top_n, percentage, threshold}

\item{cutoff_treshold}{Depending of a cutoff method, a number of features to be selected, percentage of variables to be selected, a threshold above which features are selected.}

\item{n_threads}{Number of threads for feature selection (as default set to 1)}
}
\value{
filtered set of data
}
\description{
Select variables related to a target. Features are selected based on ranking that is created depending of selected method.
Then, three methods of selection can be chosen:
}
\examples{
filtered_data <- nested_filtering(data = data_prepared, target = target, filter_name = "auc", cutoff_method = "top_n", cutoff_treshold = 10, n_threads = 10)
}
