% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_target.R
\name{prepare_data_for_modelling}
\alias{prepare_data_for_modelling}
\title{Title}
\usage{
prepare_data_for_modelling(data, target)
}
\arguments{
\item{target}{A target object defined previously}
}
\value{
List of datasets, each dataset includes target column
}
\description{
description
}
\examples{
data_with_target <- add_target(data = my_data, target = my_target)

}
